[REFACTOR] 메인페이지 폴더 구조 변경 및 오류 수정
