+++ src/components/mainpage/SwipeCard.tsx
@@ -67,7 +67,7 @@ export default function SwipeCard({
           }
         }}
         transition={{ type: "spring", stiffness: 280, damping: 24 }}
-        className={className}
+        className={`w-full ${className}`}
       >
         {/* 버튼 클릭 방해 요소 제거했습니다 */}
         {children}
