--- src/hooks/auction/useBid.ts
@@ -1,6 +1,0 @@
-import { useState } from "react";
-
-export const useBid = (initialPrice: number) => {
-  const [price, setPrice] = useState(initialPrice);
-  return { price, setPrice };
-};
