+++ src/pages/home/HomePage.tsx
@@ -1,4 +1,4 @@
-import { Logo } from "@/assets/svgs/main";
+import LoadingSpinner from "@/components/common/LoadingSpinner";
 import HeaderBell from "@/components/mainpage/HeaderBell";
 import SwipeDeck from "@/components/mainpage/SwipeDeck";
 import { useAuctionDeck } from "@/hooks/auction/useAuctionDeck";
@@ -8,26 +8,16 @@ export default function HomePage() {
   const { deck, isLoading, error, loadMore } = useAuctionDeck();
 
   return (
-    <main className="w-full">
+    <main className="relative w-full h-full bg-white">
       {/* 상단 헤더 */}
-      <header className="mx-auto w-[360px] py-4">
-        <div className="flex items-center justify-between">
-          {/* LOGO */}
-          <Logo className="w-[58px] h-auto ml-3" aria-label="LOGO" />
-
-          {/* 알림 버튼 */}
-          <div className="mr-3">
-            <HeaderBell />
-          </div>
-        </div>
-      </header>
+      <HeaderBell />
 
       {/* 카드 덱 */}
-      <section className="py-3 min-h-[640px] flex justify-center items-center">
-        {isLoading && <p className="text-bluegrey07">상품을 불러오는 중...</p>}
+      <section className="flex h-full items-center justify-center">
+        {isLoading && <LoadingSpinner />}
 
         {error && (
-          <p className="text-red-500 text-center">
+          <p className="text-center text-red-500">
             데이터를 불러오는 중 오류가 발생했습니다.
           </p>
         )}
