+++ src/types/my/myAuctionApi.type.ts
